apiserver/requirements.txt
python-dotenv==1.2.1

google-api-python-client==2.187.0
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
google-genai==1.55.0

google-auth<2.42.0

fastapi==0.101.0
uvicorn[standard]==0.23.1
